added unit tests for class names such as 'human class 1'
